[LeeCoder] Feat : 달러 원화 표기 로직 추가

+++ core/util/src/main/kotlin/com/leecoder/stockchart/util/extension/IntegerExtensions.kt
@@ -4,7 +4,8 @@ import java.text.DecimalFormat
 
 fun Double.toCurrency(): String =
     try {
-        DecimalFormat("#,###").format(this)
+        DecimalFormat("#,###").format(this.toInt())
     } catch (e: NumberFormatException) {
-        this.toString()
+        this.toInt().toString()
     }
+
